Updated Nonce, Cleaned Repo, Test script added
