--- privkeyec
@@ -1,6 +1,0 @@
------BEGIN EC PRIVATE KEY-----
-MIGkAgEBBDDMGM74OApKz6MfYi/OR+kqbHe5gQjpl2u+G2ZWtL+bUn0VEAyEI24E
-tf7WnhWIgV2gBwYFK4EEACKhZANiAARrA68nfuAjg48NDykpCahxzygJVfcBSxNP
-AZYkUtmWvQyb+adN+qVumFj96Y9+ZS19m9IFIbaMNwe84fdNxf6hWiiJD5vALa5A
-ZopxDr1mXZpA/MdGl0hGWuvI313Jono=
------END EC PRIVATE KEY-----
