--- pubkey
@@ -1,17 +1,0 @@
------BEGIN CERTIFICATE-----
-MIICqjCCAZKgAwIBAgIUWXIY+UkClUdiPxXpoNMXQIg8FjkwDQYJKoZIhvcNAQEL
-BQAwDzENMAsGA1UEAwwETm9haDAeFw0xOTEyMTQwNTA2NTZaFw0yMDEyMTMwNTA2
-NTZaMA8xDTALBgNVBAMMBE5vYWgwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEK
-AoIBAQCzBEfY2lRQyjTpjk1UNlrngbzLs86vvps/E8PWZ+TihYYaOsYYvLmmRAKX
-3tbnw5aMDqHF05MYOxVZRhnSkBpuwnaStCvdy81+U9liAUFIc6OwqTAxHvMlBCLg
-1ITirSDGWEjLZTv36uk6hN8LoVdinIy0tWG7wurtDvDpauks8nNVpXi9YqwGjJnX
-1t/WxhQfqevnT/MjZNx/VqgQZmwSYTzKm102Rx9lvdeSrwpMTULLlQRb5Eo/Hupn
-w8sRVv9ZWW6aLFi8UAgaSZSc9q6B+76ox2Ca/mwZswBKr98HLg8ASZMa57MqFx1W
-CsVbr+gJmO0Ff9eMLxd5pesL16IRAgMBAAEwDQYJKoZIhvcNAQELBQADggEBAFdh
-MgX41tdpVyvqvN0qRK5pxKmp87xnOjwa16NahacTNXTtHfRrDfhjcncWs27p+WMc
-mzqfRqdNamZhiRvqOJ2CN1q02s1zHn+x3Fl1/e24g//B1QuRMothYOTHfVwk+n8L
-qYwypf+cvW3VhWY/0upsKywEJ8ozPdgVA4A2BfA/8jPdu/wEzbXd6YEnDLyuis8S
-BddDQdZ4+T6pQbMhp4GBmXe0kXXP8Mu6ROR9vhFpj7xs0tqif4Qu60NzAfh+NLHW
-HAJwazk8rIkfdZdzAPqxfMI2wP5tsekvqxohNyufjEeEjFkAqVMfbcieah+i4KAY
-ZiS8WlowtwbIqn33o5g=
------END CERTIFICATE-----
