--- pubkeyec
@@ -1,10 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIBWjCB4aADAgECAhQ8bmJQy87Zjqv7a2LYGQGqjk1kvDAKBggqhkjOPQQDAjAP
-MQ0wCwYDVQQDDAROb2FoMB4XDTE5MTIxNDA1MDY0N1oXDTIwMTIxMzA1MDY0N1ow
-DzENMAsGA1UEAwwETm9haDB2MBAGByqGSM49AgEGBSuBBAAiA2IABGsDryd+4COD
-jw0PKSkJqHHPKAlV9wFLE08BliRS2Za9DJv5p036pW6YWP3pj35lLX2b0gUhtow3
-B7zh903F/qFaKIkPm8AtrkBminEOvWZdmkD8x0aXSEZa68jfXcmiejAKBggqhkjO
-PQQDAgNoADBlAjBtooYNFzUN6MCzfUnnXlZHRH8ResS3l7fbpuq6IETwkZet5EVh
-U74jEc3GpcS0LyMCMQCnc8fBw5395MD/h26s7kzMevS0o7K4VRAMGM+Y30qeIYoF
-YZmspA68x0ZGKZj+fhg=
------END CERTIFICATE-----
